requirements.txt
python-telegram-bot==21.4
python-telegram-bot[job-queue]
requests
pyrebase4